% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{basic_norm}
\alias{basic_norm}
\title{Normalising Constant Calculator}
\usage{
basic_norm(weights)
}
\arguments{
\item{weights}{A matrix containing all the weights from the SMC process. 
Each row represent a particle and each column represents a time-point.}
}
\value{
A single value which is the estimate of the normalising constant.
}
\description{
This function Calculates the normalising constant of a distribution from SMC samples from said distribution.
}
\details{
Where \eqn{w_t^{(j)}} represents the un-normalised weight from particle \eqn{j} and time \eqn{t}, 
this function calculates the normalising constant of the associated distribution as:
\deqn{\prod_{t=1}^T\frac{1}{N}\sum_{j=1}^N w_t^{(j)}.}
(where \eqn{T} is our final time-point and we have \eqn{N} total particles.)
}
