% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{brown_lik}
\alias{brown_lik}
\title{Brownian Motion Step Likelihood}
\usage{
brown_lik(x, x_0, t, sigma, log = F)
}
\arguments{
\item{x}{The proposed value of x}

\item{x_0}{the current value of x}

\item{t}{the time between steps}

\item{sigma}{the variance parameter}

\item{log}{Logical indicating whether the log-likelihood should be returned}
}
\value{
A single number giving the likelihood of the step
}
\description{
This gives the exact likelihood of a step in Brownian Motion conditional on the starting point
}
