% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{brown_samp}
\alias{brown_samp}
\title{Brownian Motion Step Sampler}
\usage{
brown_samp(n, x_0, t, sigma)
}
\arguments{
\item{n}{Number of points to sample}

\item{x_0}{Starting point}

\item{t}{Time between starting point and sampled point}

\item{sigma}{The value of the parameter Sigma in the Brownian Motion (see Details)}
}
\value{
A vector containing samples from the Brownian Motion
}
\description{
Generates a sample after fixed time from Brownian motion given starting point.
}
