% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_SMC.R
\name{diff_SMC}
\alias{diff_SMC}
\title{SMC Sampler 2}
\usage{
diff_SMC(
  g,
  p_delta,
  M_T_samplers,
  M_T_lik,
  n = 1000,
  timesteps = 100,
  varphi = 1,
  x_0,
  resample = T
)
}
\arguments{
\item{g}{Function which we are taking expectation over}

\item{p_delta}{true transition kernel between discrete time steps}

\item{M_T_samplers}{list of samplers for T approximating transition kernels}

\item{M_T_lik}{list of likelihood for T approximating kernels}

\item{n}{Number of particles to simulate}

\item{timesteps}{Number of time-steps to simulate}

\item{varphi}{varphi function used in approximation}

\item{x_0}{starting value}

\item{resample}{Logical indicating whether to resample or not}
}
\value{
A list with matrices containing various information from each observation, each entry gives an individual observation with 
each column being a separate time-point and reach row being a separate  particle
\item{Xmat - }{A matrix containing the simulated samples.}
\item{Wmat - }{A matrix containing the weights associated with each sample}
\item{gmat - }{A matrix containing the value of the function g evaulated at each sample.}
}
\description{
Samples using SMC as in the framework laid out in \insertCite{doucet}{SC2GroupProject}.
}
\examples{
# Set-up to process described in section 5.2 of ...
# Set our final timepoints
Final_T=1
#Set number of timesteps we will use
timesteps=20
#Get the size of each timestep
step=Final_T/timesteps

#Create sampling function
sampler <- function(x_0){
  sin_samp(1,x_0,step,pi)
}

# Create Likelihood Function
true_lik <- function(x,x_0){
  sin_lik(x,x_0,step,pi)
}

# Create g function
g <- function(x){
  zeta <- 100
  omega2 <- 1e-30
  diff=sqrt(zeta)-omega2
  if(x>-diff & x<diff){
    return(
      exp(1/(zeta-x^2))
    )
  }
  else{
    return(0)
  }
}

# Now run our SMC sampler
out <- diff_SMC(g=g,p_delta=true_lik,M_T_samplers = rep(list(sampler),timesteps),
    M_T_lik = rep(list(true_lik),timesteps), n = 1000, timesteps = timesteps,
    varphi = 0.9,x_0=5,resample = TRUE)

}
\references{
\insertRef{doucet}{SC2GroupProject}
}
