% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{ornuhl_samp}
\alias{ornuhl_samp}
\title{Ornstein-Uhlenbeck Step Sampler}
\usage{
ornuhl_samp(n, x_0, t, sigma, alpha)
}
\arguments{
\item{n}{Number of points to sample}

\item{x_0}{Starting point}

\item{t}{Time between starting point and sampled point}

\item{sigma}{The value of the parameter sigma in the Ornstein-Uhlenbeck process (see Details)}

\item{alpha}{The value of the parameter alpha in the Ornstein-Uhlenbeck process (see Details)}
}
\value{
A vector containing samples from the Ornstein-Uhlenbeck Process
}
\description{
Ornstein-Uhlenbeck Step Sampler
}
