% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sin_lik}
\alias{sin_lik}
\title{Sinusoidal Step likelihood}
\usage{
sin_lik(x, x_0, t, omega, log = F)
}
\arguments{
\item{x}{The proposed value of x}

\item{x_0}{the current value of x}

\item{t}{the time between steps}

\item{omega}{The value of the parameter omega in the Sinusoidal Brownian Motion (see Details)}

\item{log}{Logical indicating whether the log-likelihood should be returned}
}
\value{
A single number giving the likelihood of the step
}
\description{
This gives the exact likelihood of a step in Sinusoidal Brownian Motion conditional on the starting point
}
