% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sin_samp}
\alias{sin_samp}
\title{Sinusoidal Brownian Motion Step Sampler}
\usage{
sin_samp(n, x_0, t, omega)
}
\arguments{
\item{n}{Number of points to sample}

\item{x_0}{Starting point}

\item{t}{Time between starting point and sampled point}

\item{omega}{The value of the parameter omega in the Sinusoidal Brownian Motion (see Details)}
}
\value{
A vector containing samples from the Sinusoidal Brownian Motion
}
\description{
Sinusoidal Brownian Motion Step Sampler
}
