% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{basic_norm}
\alias{basic_norm}
\title{Normalising Constant Calculator}
\usage{
basic_norm(weights)
}
\arguments{
\item{weights}{A matrix containing all the weights from the SMC process. 
Each row represent a particle and each column represents a time-point.}
}
\value{
A single value which is the estimate of the normalising constant.
}
\description{
This function Calculates the normalising constant of a distribution from SMC samples from said distribution.
}
