% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_SMC.R
\name{Basic_SMC}
\alias{Basic_SMC}
\title{SMC sampler}
\usage{
Basic_SMC(
  init_sample,
  proposal_sample,
  proposal_lik,
  true_lik,
  g_lik = NULL,
  y = NULL,
  n = 1000,
  timesteps = 100
)
}
\arguments{
\item{init_sample}{sampler function for the initial distribution. 
This function takes 1 argument which is the number of samples to take}

\item{proposal_sample}{sampler function for conditional proposal}

\item{proposal_lik}{likelihood for conditional proposal}

\item{true_lik}{true likelihood (up to a constant)}

\item{g_lik}{likelihood of observed y given x}
}
\value{
A list containing the simulated particles and their associated weights
}
\description{
SMC sampler
}
