% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_SMC.R
\name{diff_SMC}
\alias{diff_SMC}
\title{Title}
\usage{
diff_SMC(
  g,
  p_delta,
  M_T_samplers,
  M_T_lik,
  n = 1000,
  timesteps = 100,
  varphi = 1,
  x_0
)
}
\arguments{
\item{g}{Function which we are taking expectation over}

\item{p_delta}{true transition kernel between discrete time steps}

\item{M_T_samplers}{list of samplers for T approximating transition kernels}

\item{M_T_lik}{list of likelihood for T approximating kernels}

\item{n}{Number of particles to simulate}

\item{timesteps}{Number of time-steps to simulate}

\item{varphi}{varphi function used in approximation}

\item{x_0}{starting value}
}
\value{
A list containing the simulated particles and their associated weights
}
\description{
Title
}
