% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_SMC_sin}
\alias{diff_SMC_sin}
\title{SMC Sampler 2}
\usage{
diff_SMC_sin(x_0, t, n, timesteps, varphi, resample)
}
\arguments{
\item{x_0}{starting value}

\item{n}{Number of particles to simulate}

\item{timesteps}{Number of time-steps to simulate}

\item{varphi}{varphi function used in approximation}

\item{resample}{Logical indicating whether to resample or not}

\item{g}{Function which we are taking expectation over}

\item{p_delta}{true transition kernel between discrete time steps}

\item{M_T_samplers}{list of samplers for T approximating transition kernels}

\item{M_T_lik}{list of likelihood for T approximating kernels}
}
\value{
A list containing the simulated particles and their associated weights
}
\description{
Samples using SMC as in the framework laid out in ...
}
