% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_SMC_sin}
\alias{diff_SMC_sin}
\title{SMC Sampler Sin}
\usage{
diff_SMC_sin(x_0, t, n, timesteps, varphi, resample)
}
\arguments{
\item{x_0}{starting value}

\item{t}{The length of each time-step.}

\item{n}{Number of particles to simulate}

\item{timesteps}{Number of time-steps to simulate (`t*timesteps` should give the end time.)}

\item{varphi}{varphi function used in approximation}

\item{resample}{Logical indicating whether to resample or not}
}
\value{
A list with matrices containing various information from each observation, each entry gives an individual observation with 
each column being a separate time-point and reach row being a separate  particle
\item{Xmat - }{A matrix containing the simulated samples.}
\item{Wmat - }{A matrix containing the weights associated with each sample}
\item{gmat - }{A matrix containing the value of the function g evaulated at each sample.}
}
\description{
Samples using SMC as in the framework laid out in \insertCite{doucet}{SC2GroupProject}.
This function uses the SMC framework to implement Example 1 of Section 5.3 in 
\insertCite{doucet}{SC2GroupProject} given in section 5.2.
}
\details{
This function carries out the \code{diff_SMC framework} for the following scenario.
The diffusion process is \eqn{\mathrm{d}X_t=\sin(X_t-\omega)\mathrm{d}_t\mathrm{d}W_t}
which we approximate with discrete time transition kernel
\deqn{p^e_{\Delta t}(x,x_0)=\sqrt{\frac{1}{2\pi \Delta t^2}}\exp\{\frac{-(x-\sin(x_0-\omega)\Delta t+x_0)^2}{2\Delta t^2}\}}
\eqn{g} is given by
\deqn{g(x)=\begin{cases}\exp\{\frac{1}{\zeta-x^2}\}\quad\text{if }x\in(\sqrt{\zeta}-\omega',\omega'-\sqrt{\zeta})\\
  0\quad\text{o/w}\end{cases}}
For \eqn{\zeta=100,~\omega'=1e^{-30}}.
}
\examples{
# Generate 10,000 samples and at 20 time-steps up to time 1
out <- diff_SMC_sin(x_0=5, t=1/20, n=10000, timesteps=20, varphi=0.9,TRUE)
}
\references{
\insertRef{doucet}{SC2GroupProject}
}
