% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_sampler.R
\name{discrete_sampler}
\alias{discrete_sampler}
\title{A function to do discrete time sampling from your distribution.}
\usage{
discrete_sampler(
  sampler,
  x_0 = NULL,
  timesteps = 100,
  n = 1000,
  init_sampler = NULL
)
}
\arguments{
\item{sampler}{The function that will be used to sample at each timestep. 
Takes in one argument which is the current value of x}

\item{x_0}{The starting point. If `NULL` then `init_sampler` must be specified.}

\item{timesteps}{The number of steps to take}

\item{n}{The number of particles to sample}

\item{init_sampler}{A function that will be used to produce the initial sample.
Take in one argument which is the number of points to sample. Only to be specified if `x_0=NULL`}
}
\value{
A matrix containing the sampled particles. Each row is a particle and each column is a time-step.
}
\description{
A function to do discrete time sampling from your distribution.
}
