% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{harmonic_norm}
\alias{harmonic_norm}
\title{Harmonic Mean Normalising Constant}
\usage{
harmonic_norm(g, weights, varphi)
}
\arguments{
\item{g}{A vector containing the final timepoint for each particle transformed by 
the function which you are trying to estimate the expectation of.}

\item{weights}{A matrix containing all the weights from the SMC process. 
Each row represent a particle and each column represents a time-point.}

\item{varphi}{The constant between 0 and 1 used in the SMC procedure}
}
\value{
A single value giving the estimated normalising constant.
}
\description{
This function calculates the normalising constant using the harmonic mean. 
This should only be used when the proposal sampler is identical to the density.
}
