% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sin_lik}
\alias{sin_lik}
\title{Sinusoidal Step likelihood}
\usage{
sin_lik(x, x_0, t, omega, log = F)
}
\arguments{
\item{x}{The proposed value of x}

\item{x_0}{the current value of x}

\item{t}{the time between steps}

\item{omega}{The value of the parameter omega in the Sinusoidal Brownian Motion (see Details)}

\item{log}{Logical indicating whether the log-likelihood should be returned}
}
\value{
A single number giving the likelihood of the step
}
\description{
This gives the approximate likelihood of a step in Sinusoidal Brownian Motion conditional on the starting point
}
\details{
The true diffusion process we are attempting to obtain likelihoods from is \eqn{\mathrm{d}X_t=\sin(X_t-\omega)\mathrm{d}t+\mathrm{d}W_t}. 
We approximate this at time `t` with transition kernel: 
\deqn{p_t(x,x_0)=\sqrt{\frac{1}{2\pi t^2}}\exp\{\frac{-(x-\sin(x_0-\omega)t+x_0)^2}{2t^2}\}}
}
