% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sin_samp}
\alias{sin_samp}
\title{Sinusoidal Brownian Motion Step Sampler}
\usage{
sin_samp(n, x_0, t, omega)
}
\arguments{
\item{n}{Number of points to sample}

\item{x_0}{Starting point}

\item{t}{Time between starting point and sampled point}

\item{omega}{The value of the parameter omega in the Sinusoidal Brownian Motion (see Details)}
}
\value{
A vector containing samples from the Sinusoidal Brownian Motion
}
\description{
This function samples an approximate step from a sinusoidal Brownian Motion.
}
\details{
The true diffusion process we are attempting to sample from is \eqn{\mathrm{d}X_t=\sin(X_t-\omega)\mathrm{d}t+\mathrm{d}W_t}. 
We approximate this at time `t` by sampling with transition kernel: 
\deqn{p_t(x,x_0)=\sqrt{\frac{1}{2\pi t^2}}\exp\{\frac{-(x-\sin(x_0-\omega)*t+x_0)^2}{2t^2}\}}
}
